+++ Cabal/Distribution/Simple/InstallDirs.hs
@@ -395,7 +395,10 @@ type PathTemplateEnv = [(PathTemplateVariable, PathTemplate)]
 -- | Convert a 'FilePath' to a 'PathTemplate' including any template vars.
 --
 toPathTemplate :: FilePath -> PathTemplate
-toPathTemplate = PathTemplate . read -- TODO: eradicateNoParse
+toPathTemplate fp = PathTemplate
+    . fromMaybe (error $ "panic! toPathTemplate " ++ show fp)
+    . readMaybe -- TODO: eradicateNoParse
+    $ fp
 
 -- | Convert back to a path, any remaining vars are included
 --